edit search input

+++ app/Http/Controllers/Evaluation/EvaluationController.php
@@ -416,20 +416,23 @@ class EvaluationController extends Controller
                 return response()->json(['status'=> 'success', 'data' => $id]);
                 break;
 
-            case 'deleteCreateEvaluation':
-                $id          = $request->get('id');
-                //sd($id);
-                $id_topic    = CreateEvaluation::with('parts', 'parts.question')->where('id_topic', $id)->first();
-                //sd($id_topic->toArray());
-                //sd($id_question->toArray());
-                $id_topic ->delete();
-                return response()->json(['status'=> 'success', 'data' => $id]);
-                break;
-
             default:
                 # code...
                 break;
         }
+    }
 
+    public function postDeleteCreateEvaluation($id)
+    {
+        $id_topic    = CreateEvaluation::with('parts', 'parts.question')->where('id_topic', $id)->first();
+
+        if(!empty($id_topic)){
+
+            $id_topic->delete();
+
+            return['status'     => 'success', 'message' => 'Delete complete.'];
+        } else {
+            return['status'     => 'failed','message'   =>'Not found.'];
+        }
     }
 }
